refactor(instruction): rename fiber to instruction
